refactor scale utils
closes #437
