+++ projects/ngx-d3timeline/src/lib/scales/scale-utils.ts
@@ -1,9 +1,4 @@
-import { TimeScale, BandScale, InverseBandScale } from './scale-types';
-import { Orientation } from '../core/orientation';
-import { scaleBand, scaleTime } from 'd3-scale';
-import { Activity } from '../activity/activity';
-import { min, max } from 'd3-array';
-import { View } from '../view/view';
+import { BandScale, InverseBandScale } from './scale-types';
 
 export function getInverseBandScale(scale: BandScale): InverseBandScale {
   const domain = scale.domain();
@@ -13,62 +8,4 @@ export function getInverseBandScale(scale: BandScale): InverseBandScale {
     const index = Math.floor((value - paddingOuter) / eachBand);
     return domain[Math.max(0, Math.min(index, domain.length - 1))];
   };
-}
-
-export function rescaleTime(
-  activities: Activity[],
-  view: View,
-  timeOrientation: Orientation,
-  event: any
-): TimeScale {
-  const unscaledTimeScale = configureTimeScale(
-    activities,
-    view,
-    timeOrientation
-  );
-
-  return event
-    ? timeOrientation === Orientation.Vertical
-      ? event.transform.rescaleY(unscaledTimeScale)
-      : event.transform.rescaleX(unscaledTimeScale)
-    : unscaledTimeScale;
-}
-
-export function configureBandScale(
-  activities: Activity[],
-  view: View,
-  orientation: Orientation,
-  resourceGap: number
-): BandScale {
-  return scaleBand()
-    .domain(getBandScaleDomain(activities))
-    .range(getRange(view, orientation))
-    .paddingInner(resourceGap);
-}
-
-export function configureTimeScale(
-  activities: Activity[],
-  view: View,
-  orientation: Orientation
-): TimeScale {
-  return scaleTime()
-    .domain(getTimeScaleDomain(activities))
-    .range(getRange(view, orientation));
-}
-
-function getBandScaleDomain(activities: Activity[]): string[] {
-  return [...new Set(activities.map(activity => activity.resource))];
-}
-
-function getTimeScaleDomain(activities: Activity[]): [Date, Date] {
-  return [
-    min(activities, activity => activity.start),
-    max(activities, activity => activity.finish)
-  ];
-}
-
-function getRange(view: View, orientation: Orientation): [number, number] {
-  return orientation === Orientation.Vertical
-    ? [view.top, view.bottom]
-    : [view.left, view.right];
 }
